--- a/index.mjs
+++ b/index.mjs
@@ -1,5 +1,18 @@
 import parse from 'regexparam';
 
+const mthMap = {
+	GET: 1,
+	HEAD: 2,
+	PATCH: 3,
+	OPTIONS: 4,
+	CONNECT: 5,
+	DELETE: 6,
+	TRACE: 7,
+	POST: 8,
+	PUT: 9,
+	"": 0, // 'all'
+}
+
 export default class Trouter {
 	constructor() {
 		this.routes = [];
@@ -19,24 +32,25 @@ export default class Trouter {
 	use(route, ...fns) {
 		let handlers = [].concat.apply([], fns);
 		let { keys, pattern } = parse(route, true);
-		this.routes.push({ keys, pattern, method:'', handlers });
+		this.routes.push({ keys, pattern, method: '', handlers, mthidx: mthMap[''] });
 		return this;
 	}
 
 	add(method, route, ...fns) {
 		let { keys, pattern } = parse(route);
 		let handlers = [].concat.apply([], fns);
-		this.routes.push({ keys, pattern, method, handlers });
+		this.routes.push({ keys, pattern, method, handlers, mthidx: mthMap[method] });
 		return this;
 	}
 
 	find(method, url) {
-		let isHEAD=(method === 'HEAD');
+		let mthidx = mthMap[method];
+		let isHEAD = (mthidx === 2);
 		let i=0, j=0, k, tmp, arr=this.routes;
 		let matches=[], params={}, handlers=[];
 		for (; i < arr.length; i++) {
 			tmp = arr[i];
-			if (tmp.method.length === 0 || tmp.method === method || isHEAD && tmp.method === 'GET') {
+			if (tmp.mthidx === mthidx  || !tmp.mthidx || (isHEAD && tmp.mthidx===1) ) {
 				if (tmp.keys === false) {
 					matches = tmp.pattern.exec(url);
 					if (matches === null) continue;