It's saving ~10ms

closes #19