@Kikobeats 2023-02-20
done; next PR will be to move from Travis to GitHub Actions 😛